# knowledge/adapters/openapi.yaml
openapi: 3.1.0
info:
  title: advertisment_system inference contract
  version: "1.0"
  description: |
    The three endpoints every media service (image, video, tts) must speak.
    Prose and per-kind field meanings: CONTRACT.md. LLMs are OpenAI-compatible
    servers and are not described here.
paths:
...
  /generate:
    post:
      summary: run one generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/ImageRequest"
                - $ref: "#/components/schemas/VideoRequest"
                - $ref: "#/components/schemas/TtsRequest"
      responses:
        "200":
          description: files are relative to the service's ASSETS_DIR
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ImageResponse"
                  - $ref: "#/components/schemas/VideoResponse"
                  - $ref: "#/components/schemas/TtsResponse"
        "422":
          description: unsupported model / bad request — never retried by the worker
        "5xx":
          description: transient; the worker marks the endpoint unhealthy and fails over
  /files/{path}:
    get:
      summary: fetch a generated file
      parameters:
        - in: path
          name: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: the bytes }
        "404": { description: unknown, or resolves outside ASSETS_DIR }
components:
  schemas:
...
    ImageRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string }
        product_id: { type: string, default: adhoc }
        kind: { type: string, default: image }
        n: { type: integer, default: 1 }
        width: { type: integer, default: 1024 }
        height: { type: integer, default: 1024 }
        steps: { type: integer, nullable: true }
        seed: { type: integer, nullable: true }
        lora_name: { type: string, nullable: true }
        lora_strength: { type: number, nullable: true }
        lora_base: { type: string, nullable: true, description: The rung the LoRA was trained on; the host skips the LoRA when rendering another rung. }
        model: { type: string, nullable: true, description: A rung from ImageCaps.models; absent = the host default. }
        ref_image_path: { type: string, nullable: true, description: Edit rungs — the photo to edit, relative to the host's assets dir. }
        ref_image_b64: { type: string, nullable: true, description: Edit rungs — the photo as base64 for a worker on another machine; wins over ref_image_path. }
        ref_fit: { type: string, nullable: true, enum: [contain], description: Edit rungs — fit the whole photo into width x height (blurred fill) so the output has the requested frame; absent = output follows the reference. }
    VideoRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string }
        product_id: { type: string, default: adhoc }
        ref_image_path: { type: string, nullable: true }
        ref_image_b64:
          type: string
          nullable: true
          description: >
            The reference image as base64 bytes, for a worker on another machine whose
            assets this host cannot read. Wins over ref_image_path when both are given.
        quality: { type: string, default: draft }
        width: { type: integer }
        height: { type: integer }
        num_frames: { type: integer }
        seed: { type: integer, nullable: true }
        model: { type: string, enum: [5b, 14b, ltx2], nullable: true }
    TtsRequest:
      type: object
      required: [text]
      properties:
        text: { type: string }
        product_id: { type: string, default: adhoc }
        lang: { type: string, default: en }
        voice: { type: string, nullable: true }
        pronunciations: { type: object, nullable: true }
        ref_audio_path: { type: string, nullable: true }
    ImageResponse:
      type: object
      required: [images]
      properties:
        model: { type: string }
        images:
          type: array
          items:
            type: object
            required: [path]
            properties:
              path: { type: string, description: relative to ASSETS_DIR }
              seed: { type: integer }
              elapsed_s: { type: number }
    VideoResponse:
      type: object
      required: [video]
      properties:
        video:
          type: object
          required: [path]
          properties:
            path: { type: string }
            elapsed_s: { type: number }
            frames: { type: integer }
            model: { type: string }
            quality: { type: string }
    TtsResponse:
      type: object
      required: [audio]
      properties:
        model: { type: string }
        engine: { type: string }
        audio:
          type: object
          required: [path]
          properties:
            path: { type: string }
            sample_rate: { type: integer }
            duration_s: { type: number }
            elapsed_s: { type: number }